requirements.txt
Django==2.1.7
djangorestframework==3.9.2
django-url-filter==0.3.12
django-filter==2.1.0
docker==3.7.0
psycopg2==2.7.7
djangorestframework-bulk==0.2.1